tests/wpt/tests/dom/nodes/Document-contentType/contentType/contenttype_bmp.html

<!DOCTYPE html>
<title>BMP document.contentType === 'image/bmp'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var iframe = document.createElement('iframe');
  iframe.addEventListener('load', this.step_func_done(function() {
    assert_equals(iframe.contentDocument.contentType, "image/bmp");
  }), false);
  iframe.src = "../resources/t.bmp";
  document.body.appendChild(iframe);
});
</script>
