w3c-reports/12092014/gtk/runned-tests/dom/nodes/Document-contentType/contentType/contenttype_bmp-expected.txt


FAIL BMP document.contentType === 'image/bmp' assert_equals: expected (string) "image/bmp" but got (undefined) undefined(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/dom/nodes/Document-contentType/contentType/contenttype_bmp.html:11:18
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:843:37)
